requirements.txt
# Основные зависимости
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Дополнительные платформы
discord.py==2.3.2
vk-api==11.9.9

# Утилиты
requests==2.31.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0